pyproject.toml
[project]
name = "bistro-e2e-python"
version = "1.0.0"
description = "E2E tests for Bistro Delivery using Playwright for Python"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.50",
    "pytest>=8.0",
    "pytest-playwright>=0.7",
    "pydantic>=2.12",
    "faker>=33.0",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--screenshot=on",
    "--tracing=retain-on-failure",
    "--junitxml=junit-results/results.xml",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
